chore: dependency vendoring of `markdownlint-rule-helpers`
