+++ package.json
@@ -44,27 +44,25 @@
     "postpublish": "pinst --enable"
   },
   "dependencies": {
-    "markdown-it": "14.0.0",
-    "markdownlint-rule-helpers": "0.24.0"
+    "markdown-it": "14.0.0"
   },
   "devDependencies": {
     "@commitlint/cli": "18.4.4",
     "@commitlint/config-conventional": "18.4.4",
     "@types/markdown-it": "13.0.7",
-    "@types/node": "20.10.8",
+    "@types/node": "20.11.0",
     "editorconfig-checker": "5.1.2",
     "eslint": "8.56.0",
     "eslint-config-conventions": "13.1.0",
     "eslint-config-prettier": "9.1.0",
     "eslint-plugin-import": "2.29.1",
-    "eslint-plugin-prettier": "5.1.2",
+    "eslint-plugin-prettier": "5.1.3",
     "eslint-plugin-promise": "6.1.1",
     "eslint-plugin-unicorn": "50.0.1",
     "husky": "8.0.3",
     "lint-staged": "15.2.0",
     "markdownlint": "0.33.0",
     "markdownlint-cli2": "0.11.0",
-    "markdownlint-rule-helpers": "0.24.0",
     "pinst": "3.0.0",
     "prettier": "3.1.1",
     "semantic-release": "22.0.12",
